Exporting platform_impl_lib headers (#8045)
This allows consuming projects using repository overlaying to disambiguate overlapping include paths when it comes to platform_impl.h by going through envoy/external/...

Addendum to #8005

Risk Level: Low
Testing: N/A
