+++ source/exe/BUILD
@@ -57,7 +57,8 @@ envoy_cc_library(
     ],
     deps = [
         ":envoy_main_common_lib",
-    ] + envoy_cc_platform_dep("platform_impl_lib"),
+        ":platform_impl_lib",
+    ],
 )
 
 envoy_cc_library(
@@ -66,6 +67,7 @@ envoy_cc_library(
     hdrs = ["main_common.h"],
     deps = [
         ":envoy_common_lib",
+        ":platform_impl_lib",
         ":process_wide_lib",
         "//source/common/api:os_sys_calls_lib",
         "//source/common/common:compiler_requirements_lib",
@@ -80,7 +82,7 @@ envoy_cc_library(
             "//source/common/signal:sigaction_lib",
             ":terminate_handler_lib",
         ],
-    }) + envoy_cc_platform_dep("platform_impl_lib"),
+    }),
 )
 
 envoy_cc_library(
@@ -96,11 +98,26 @@ envoy_cc_library(
     ] + envoy_google_grpc_external_deps(),
 )
 
+envoy_cc_library(
+    name = "platform_impl_lib",
+    deps = [":platform_header_lib"] +
+           envoy_cc_platform_dep("platform_impl_lib"),
+)
+
+envoy_cc_library(
+    name = "platform_header_lib",
+    hdrs = ["platform_impl.h"],
+    deps = [
+        "//include/envoy/filesystem:filesystem_interface",
+        "//include/envoy/thread:thread_interface",
+    ],
+)
+
 envoy_cc_posix_library(
     name = "platform_impl_lib",
-    hdrs = ["posix/platform_impl.h"],
-    strip_include_prefix = "posix",
+    srcs = ["posix/platform_impl.cc"],
     deps = [
+        ":platform_header_lib",
         "//source/common/common:thread_lib",
         "//source/common/filesystem:filesystem_lib",
     ],
@@ -108,9 +125,9 @@ envoy_cc_posix_library(
 
 envoy_cc_win32_library(
     name = "platform_impl_lib",
-    hdrs = ["win32/platform_impl.h"],
-    strip_include_prefix = "win32",
+    srcs = ["win32/platform_impl.cc"],
     deps = [
+        ":platform_header_lib",
         "//source/common/common:assert_lib",
         "//source/common/common:thread_lib",
         "//source/common/filesystem:filesystem_lib",
